Add changes to fix one of the tests
Fixes regex that is only simple and not complex at all.
Currently since we can't control when TOP is ran, we put
the simple license we find into @!array if it's currently empty
because we will need itto end up in @!all-licenses in the end
(it won't end up anywhere unless it is pushed to @!array)
Since we don't have a method for simple ones, this will work for now.